code queries: finalizing code for loading hbase and get friends and mvPOI queries

+++ queries/src/main/java/hbase_schema/GPStracesTable.java
@@ -15,7 +15,6 @@ import org.apache.hadoop.hbase.client.Put;
 import org.apache.hadoop.hbase.client.Result;
 import org.apache.hadoop.hbase.util.Bytes;
 
-import containers.CheckIn;
 import containers.GPStrace;
 import parser.ParseGPStrace;
 
@@ -23,22 +22,25 @@ public class GPStracesTable implements QueriesTable {
 
 	protected String tableName;
 	protected HTable table;
+	private int regionsNo, usersNo;
 
-	public GPStracesTable(String nm) {
+	public GPStracesTable(String nm, String u, String r) {
 		this.tableName = nm;
+		this.usersNo = Integer.parseInt(u);
+		this.regionsNo = Integer.parseInt(r);
 	}
 
 	public void createTable() throws Exception {
 		Configuration hbaseConf = HBaseConfiguration.create();
 		HBaseAdmin admin = new HBaseAdmin(hbaseConf);
-		/*if (admin.tableExists(this.tableName)) {
+		if (admin.tableExists(this.tableName)) {
 			admin.disableTable(this.tableName);
 			admin.deleteTable(this.tableName);
-		}*/
+		}
 		HTableDescriptor descriptor = new HTableDescriptor(this.tableName);
 		descriptor.addFamily(new HColumnDescriptor("gpsTraces"));
 
-		//admin.createTable(descriptor);
+		admin.createTable(descriptor, this.getSplitKeys());
 		admin.close();
 		this.table = new HTable(hbaseConf, this.tableName);
 	}
@@ -62,6 +64,18 @@ public class GPStracesTable implements QueriesTable {
 			System.out.println("...single gps-trace retrieved.");
 		}
 	}
+	
+	public byte[][] getSplitKeys() {
+		int keyNo = usersNo / regionsNo;
+		byte[][] keys = new byte[regionsNo][Integer.SIZE];
+
+		int c = 1;
+		for (int i = 0; i < regionsNo; i++) {
+			keys[i] = Bytes.toBytes(c);
+			c += keyNo;
+		}
+		return keys;
+	}
 
 	public static void main(String[] args) throws Exception {
 
@@ -69,7 +83,7 @@ public class GPStracesTable implements QueriesTable {
 		GPStrace tr;
 		byte[] row, qualifier, data;
 		
-		GPStracesTable trTable = new GPStracesTable("gps-traces");
+		GPStracesTable trTable = new GPStracesTable("gps-traces", args[1], args[2]);
 		System.out.println("Creating HBase gps-traces table...");
 		trTable.createTable();
 		System.out.println("...done");
@@ -79,8 +93,7 @@ public class GPStracesTable implements QueriesTable {
 		BufferedReader br = new BufferedReader(fr);
 
 		line = br.readLine();
-		while (line != null){
-			
+		while (line != null){	
 			if (!line.equals(" ")) {
 				tr = pr.parseLine(line);
 				tr.print();
@@ -88,11 +101,11 @@ public class GPStracesTable implements QueriesTable {
 				qualifier = tr.getQualifierBytes();
 				row = tr.getKeyBytes();
 				trTable.putSingle(row, qualifier, data);
-				//trTable.getSingle(row, qualifier);
+				trTable.getSingle(row, qualifier);
 			}
 			line = br.readLine();
 		}
-		
+		br.close();
 	}
 
 }
